dibuja() reina

+++ src/tablero.h
@@ -5,11 +5,11 @@ class Tablero {
 
 	int filas;
 	int columnas;
-	float casillaSize;
+	double casillaSize;
 
 public:
 
-	Tablero(int f = 8, int c = 8, float size = 1.5f) : filas(f), columnas(c), casillaSize(size) {}
+	Tablero(int f = 8, int c = 8, float size = 1.5) : filas(f), columnas(c), casillaSize(size) {}
 
 	void dibujarTablero();
